hankblog

+++ src/main/java/com/blog/Repository/CurrencyRepository.java
@@ -4,8 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.blog.pojo.Currency;
 
-
-public interface CurrencyRepository extends JpaRepository<Currency, Long> {
+public interface CurrencyRepository extends JpaRepository<Currency, Integer> {
 	
 	
 	Currency findByCode(String code);
